Move resizeImageToNextPowerOfTwo to separate file
